adjust Solana.RPC.Middleware to decode responses
oftentimes RPC API responses have base58- or base64-encoded fields. We
can decode them since we know which request corresponds to each
response, so we know the structure of each response. This means we won't
need to decode responses when calling the top-level module functions, so
we can feed them directly into other API request geneators (which assume
the input arguments aren't encoded).

For now, this only decodes responses for:

- requestAirdrop
- sendTransaction
- getRecentBlockhash