add Transaction.parse/1, improve CompactArray ergonomics
this commit adds Transaction.parse/1, which parses a Transaction
(without the signers) out of an encoded binary, i.e. one created with
Transaction.to_binary/1. This functionality was previously in place in
the TransactionTest module, so remove it there as well.

Additionally, allow CompactArray decoding to return an error if the
array is invalid, since we now expose this functionality publicly.