Fix invalid port range, deprecate RPC method (#4)
`solana-test-validator --help` (version 2.2.21) states that the default
dynamic port range defaults to `1024-65535`. However, passing this value
results into an `"Invalid dynamic_port_range"` error.

According to the
[source code](https://github.com/solana-labs/solana/blob/master/validator/src/cli.rs#L2207),
the default value for dynamic_port_range is actually `8000-10000`. Thus,
this commit updates the default for this field in `Solana.TestValidator`
to reflect this change, which eliminates the above error.

Additionally, a couple tests failed because they used the RPC method
`getRecentBlockhash`, which is deprecated. This commit adds support for
its replacement, `getLatestBlockhash`, and modifies the tests to use
this method instead.

Finally, this commit updates other project dependencies to resolve all
warnings on project compilation.